+++ spec/spec_helper.rb
@@ -1,5 +1,7 @@
 require 'bundler/setup'
 
+require 'byebug'
+
 require 'simplecov'
 SimpleCov.start
 SimpleCov.minimum_coverage_by_file 90