+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## [0.1.0]
+* Users Examples 1 - 5 https://docs.microsoft.com/en-us/graph/api/user-list
 * Initial Version